dcapstone/requirements.txt
# local package
-e .

python package
sklearn package

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
python3